Fix grammar details (#12)
* Inline `CHARACTER` token in `CHAR_LITERAL`

Before this, the lexer could not know whether a single-character such as
`a` should be an `IDENT` or a `CHARACTER` token. Internal priority rules
applied, which preferred `CHARACTER` over `IDENT`, leading to incorrect
parses (e.g. of `a;`).

This commit solves this by inlining `CHARACTER` in `CHAR_LITERAL`. This
is possible, because currently it is the only place where `CHARACTER` is
used. Now the lexer can distinguish them, because `CHAR_LITERAL`s have
to start with a single quote (`'`).

* Add EOF token to grammar end

* Fix grammar to allow zero-prefixed hexadecimal literals